[MT-1702] new format of SDKSettings (#302)
* [MT-1702] new format of SDKSettings

* update tests, add converters to Core and Module Settings

* Fix Dictionary Equality

NSDictionary.isEqual only works with NSObjects, so we can't use it to compare Dictionaries of ModulesSettings.

* fix failing tests

* add deep merge tests, fix Settings merge depth level

* Use subscript instead of updateValue

We don't need the return value of updateValue

* Simplify DeepMergeTests

* Improve DispatchManager and QueueManager tests

* Improve Modules and Managers Tests

* Improve lifecycle tests and fix updateSettings method definition

* Improve DeepMerge tests

* Improve SettingsManager+MergeTests

* Unify sdkSettings publish in a didSet

* Polishing DataObject usages and add deepMerge documentation

* Improve not modified logs

Avoid logging an error when resource is not modified

* Improve some TealiumModuleFactory documentation

* [MT-1702] Rename ModuleSettings.settings from Settings to Configuration (#303)

* Rename module settings to configuration

ModuleSettings contain a configuration DataObject.
That object, and the models created by with that, and the properties used to hold them,
are being renamed to configuration, to differentiate with the bigger ModuleSettings object
and to make it clear which one is required in each place.

---------

Co-authored-by: Enricoza <enricozannini93@gmail.com>