[MT-1705] Mapping Transformation (#309)
* Create Mapping and add them to the ModuleSettings and ModuleSettingsBuilder

* Create JSONTransformation for mappings and use it in the TransformerCoordinator

* Move the Mappings parsing outside of the TransformerCoordinator

* Create DataObjectConvertible protocol and split files

* Make JSONOperation generic per different Parameters

* Fix BUG IN CLIENT OF libsqlite3.dylib warning

DatabaseProvider reutilization without deallocating the previous one was duplicating the connection to the same DB causing the issue.

* Create a new DataObject init to automatically remove nils in the root

Test ModuleSettingsBuilder setMappings to generate objects correctly
Make VariableWrapper and MappingParameters only generate keys for which the values are not nil.

* Improve usage of DataObject

* Move and rename Json Transformation Configiuration objects

* Add documentation and some other rename

Remove ValueContainer possibility to contain arrays in order to simplify.

* Add register/unregister tests

* Add tests for ValueContainer and VariableAccessor

* Add Transformation operations tests

* Add test for transformation of mappings to TransformationSettings

* Fix Trace tests

* Add review improvements

Rename ScopedBarriers CodingKeys to Keys
Move jsonTransformerId to Transformers namespace
Make JsonTransformationConfiguration Generic and use the TransformationOperation type in the array
Make mappings in ModuleSettings be typed, which increases clarity even if it will require one 2 extra parsing to DataObject and back.
Fix some documentation

* Complete rename of operationsType and add tests for TealiumConfig transformations creation

* Fix Condition Converter to be compacting the input Dictionary