[MT-1708] Barrier Configuration (#316)
* Rename ScopedBarrier to BarrierSettings

Move BarrierSettings in SDKSettings
Add configuration to BarrierSettings

* Create Barrier Manager and use it in the Coordinator

* Make BarrierManager update scopes based on settings

* Add new BarrierCoordinatorTests

* Change DeepLinkHandlerWrapperTests to only handle after subscription

To avoid subscription being too slow and missing the onDataUpdated event

* Update ConnectivityBarrier and create Factory and Settings

Use the Barrier as a default with Collect.

* Make ConnectivityManager less reliant on EmpiricalConnectivity (#318)

Make the ConnectivityManager use the EmpiricalConnectivity ONLY to validate when connection is actually down.
It won't use the EmpiricalConnectivity "debounce" to assume that connection came back if monitoredConnection is still notConnected.

* Updated Documentation and improved tests

Renamed some properties to make them more readable

* Update some documentation