[MT-1601] Finalize Tealium (#341)
* Create ExecutorProxy and make ModuleProxy and Tealium use that

Made subscribe return discardable for Single
Make tealium deinit capture proxy instead of onTealiumImplementation observable (untested)

ExecutorProxy is a separate class, and is added as a dependency for ModuleProxy instead of inheritance,
because we wanted to have a more specific method name, executeModuleTask, for the module proxies
And wanted to use Task and AsyncTask with specific names (module) instead of executor.
Having ExecutorProxy being inherited would've made it's Task and AsyncTask specification, as well as it's methods
with different names (like executeTask) being visible on the ModuleProxy.

Tealium.track for now returns Result<TrackResult, Error>.
Might be too complicated, and we could simplify by mapping to just one TrackResult like this:

        let dispatch = Dispatch(name: name, type: type, data: data)
        let single: any Single<Result<TrackResult, Error>> = proxy.executeAsyncTask { tealium, completion in
            tealium.track(dispatch) { result in
                completion(.success(result))
            }
        }
        return single.asObservable().map { result in
            switch result {
            case .success(let trackResult):
                return trackResult
            case .failure(let error):
                return TrackResult(dispatch: dispatch,
                                   status: .dropped,
                                   info: "An error occurred: \(error)")
            }
        }.asSingle(queue: proxy.queue)

For now it's left with the double result instead.
There's probably a more efficient way to do this anyway, but would require some more changes probably.
Maybe a mapper function could be added.

* Make Tealium and TealiumInstanceManager use the same queue

In test provide the main queue to avoid going async.

* Create a SingleResult type alias to avoid specifying the generic error all the time

In future we could extend this to have the extra argument as well
and therefore provide more specific errors, but for now this will do.

* Rename ExecutorProxy to a more generic AsyncProxy

Add some documentation

* Remove moduleNotEnabled in favour of the generic and more descriptive `objectNotFound`

Improved TealiumError case signatures and add documentation

* Reduce swiftlint directives

* Improve Tealium documentation

* Fix indentation

* Make clearer that some of the module wrappers will be generated every time they get called

Idea here is to differenciate from modules like DataLayer where the user can just call it how many times it wants without affecting performances.
With lifecycle, instead, we don't cache it so using a function makes the difference with DataLayer, and similar methods, more prominent.

* Fix typos and use single instead of composing operators manually in AsyncProxy

* Convert flushEventQueue to return a Single and use the proxy to execute the task

* Remove Tealium prefix from TealiumModule and TealiumModuleFactory (#343)

* Remove Tealium prefix from TealiumModule and TealiumModuleFactory

Generic Specialization have a `Specific` prefix now

* Remove typealias Module to avoid confusion with the Module protocol