[MT-1599] Increase Code Coverage (#346)
* Remove AnyCodable equality code from the repo and move it in the tests

Add more tests

* Reduce places where the TealiumQueue.worker is hardcoded

There are still a few places, but there should be only 1 (or close to that) to allow us to safely pass another one in tests
Add Condition Tests
Avoid SQLite bug when deleting DB file from the wrong queue

* Add deeplink wrapper tests and pass observer directly to AsyncTask.

* Add more Condition tests and reuse VariableAccessor API to build those Conditions.

* Use xcode 16.2 version in CICD

* Use macos 15 and make Lifecycle end to end test only run on iOS

* Update fastlane and select newer iPhone for tests

* [MT-1599] Tealium+Collect end to end tests (#347)

* Make HTTPClient configurable for testing

Add some Collect end to end testing

* Create MockModuleFactory and use it to test Tealium with transformers

Improve MockBarrierFactory to always be able to access the internal barrier.

* Add dispatchPrecondition to make sure collect sends events from the right queue

* Add test for collected data

* Update CMPAdapter to return a Set of purposes and improve ConsentView

* Make ConsentDecision use a set for purposes as well

* Add test plans for all schemes to avoid parallelization

* Rename Modules to {Name}Module (#348)

* Rename Modules to {Name}Module

This is to standardize the names, make it clear when something is a module, and distinguish them from the Wrapper.
Having always Module, and not Collector/Dispatcher, makes it easier to find them.

Add Data suffix to module names that return data, like TealiumData and ConnectivityData and TimeData

* Rename DeepLink interface back to DeepLinkHandler

* Rename {ModuleName}Configuration to {ModuleName}ModuleConfiguration

* Fix Rename Tests

Fix lint warnings

* Fix DeepLink name in comments

* Revert DeepLinkHandlerWrapper rename and its tests

Rename CollectModule tests
Rename TealiumCollector directory in tests