[MT-1791] Background Task: Fix flush (#353)
* Make sure sleep events are sent on backgrounding in the same batch as any previous batched events

* Start a background task when app is going to background to allow to flush

When the task is cancelled, we stop the flush, to avoid writing to disk while the app is being suspended
When flushing we take for a limited amount of time, if the queue is not flushed after 5 seconds we stop the flush anyway.

* Add a test for lifecycle events being batched in the current flush

* Start background task only once

Add test that background task is started (only the boolean in the subject)

* Make sure to publish ongoingBackgroundTask only if changed and only from one place.

* Extract BackgroundTaskStarter class and make it thread safe

Add more tests

* Fix spelling

* Use in memory DB in tests

* Specify timeouts for BackgroundTaskStarter tests

* Use CustomObservable, not Callback observable

Remove Callback(fromDisposable) as it was exactly the same as CustomObservable under the hood.

* Try to increase fastlane xcodebuild timeout with an env

* Specify FASTLANE_XCODEBUILD_SETTINGS_TIMEOUT env directly in the script

* Create default config in tests with databaseName and networkClient already in them

* Raise QoS of TealiumQueue while testing it

* Make ApplicationStatusListenerTests more reliable by using a 0 nanoseconds leeway

* Refactor before and after methods of TimeFrame not to be optional

* Make Dispatch lazy to avoid being created with an old timestamp

The SessionManagerTests class (as any XCTestCase class) is created at the start of the tests N times (one per test method).
And then only later is used, when it's their turn to run (because we are running them in series).
This meant that the tests could fail if it took more than 5 minutes to reach the SessionManagerTests
because the Dispatch had more than 5 minutes delay from the current time, and the session would result expired already
Since it had a 5 minute timeout.

* Improve ApplicationStatusListenerTests

Use a high priority queue

* Make TimeFrame safer with big Integers

Avoid crashing due to overflows in rare cases where the numbers are too big
Use TimeFrame always instead of using Double sometimes
Remove old `addSeconds` and `addMinutes` methods since we have now unfailable TimeFrame alternatives.

* Make SQLiteOpener and DatabaseHelper take TealiumConfig alone without databaseName