[MT-1858] JSONPath (#381)
* Accessor POC

* Rename Accessor to JSONPath and fix description to match parser input

* First implementation of the JSONPathParser

* Add some tests

* Create a private, non copyable and consuming on start, Parsing struct

This is done to avoid passing the pathString and cursor to every method
Making it non copyable and consuming on start makes it so that it can only be used once, so the cursor can't be in a wrong state from a previous usage.

* Improve alphanumeric validation

* Add some more tests

* Make JSONPathParserError.ErrorType private

To avoid possible future breaking changes
Add documentation

* Refactor JSONPathParser code for clarity and add some comments

* Improve error handling and fix edge cases like emptyPathComponent

* Split in different files

* Split JSONPath files and add documentation

* Add a couple more tests for the json path parsing

* Fix parse documentation

* Update docs and use static regex.

It should never fail anyway so the error is not very important.
And doing it multiple times doesn't change the behavior.

* Use typed throws for JSONPathParser

XCode 16 introduced the ability to use swift 6 typed throws out of the box

* Rename ErrorType to Kind to follow swift suggested namings in documentation

* Improve docs and remove unnecessary sample code

* Unify JSONPath and JSONPathParser

Improve tests and fix description for invalid characters

* Reduce amounts of errors and code duplication

* Fix documentation comments

* Fix typo

* Move JSONPath to Data folder